수정 삭제 기능 수정

+++ app/controller/posts/ctrl.posts.js
@@ -100,22 +100,29 @@ const postProcess = {
       const { postId } = req.params;
       const { songName, desc, singer, url, category1, category2, category3 } =
         req.body;
-
-      await Post.update(
-        {
-          songName,
-          desc,
-          singer,
-          url,
-          category1,
-          category2,
-          category3,
-        },
-        {
-          where: { postId: postId },
-        }
-      );
-      res.send({ ok: true, message: "게시글을 수정했습니다" });
+      const post = await Post.findByPk(postId);
+      if (userId == post.userId) {
+        await Post.update(
+          {
+            songName,
+            desc,
+            singer,
+            url,
+            category1,
+            category2,
+            category3,
+          },
+          {
+            where: { postId: postId },
+          }
+        );
+        res.send({ ok: true, message: "게시글을 수정했습니다" });
+      } else {
+        res.send({
+          ok: false,
+          message: "작성자만 사용 가능한 기능입니다.",
+        });
+      }
     } catch (err) {
       res.status(400).send({
         ok: false,
@@ -129,17 +136,22 @@ const postProcess = {
     try {
       // 로그인 유저 확인
       const { userId } = res.locals.targetUserInfo;
-      console.log("userId:", userId);
       const { postId } = req.params;
       const post = await Post.findByPk(postId);
-      await post.destroy({
-        where: { postId: postId },
-      });
-
-      res.send({
-        ok: true,
-        message: "게시글을 삭제했습니다.",
-      });
+      if (userId == post.userId) {
+        await post.destroy({
+          where: { postId: postId },
+        });
+        res.send({
+          ok: true,
+          message: "게시글을 삭제했습니다.",
+        });
+      } else {
+        res.send({
+          ok: false,
+          message: "작성자만 사용 가능한 기능입니다.",
+        });
+      }
     } catch (err) {
       res.status(400).send({
         ok: false,
